Stub out new Grouping & Aggregation methods, implement Group by Index
Introduces a new datum converter for grouped data,
DictionaryDatumConverterFactory, which only supports $reql_type$
GROUPED_DATA, not generic IDictionary support.  Some more of the old
grouping support is commented out with #if...#endif until all new
grouping support is completed.

+++ rethinkdb-net/QueryTerm/CountQuery.cs
@@ -3,7 +3,7 @@ using System;
 
 namespace RethinkDb.QueryTerm
 {
-    public class CountQuery<T> : ISingleObjectQuery<double>
+    public class CountQuery<T> : ISingleObjectQuery<int>
     {
         private readonly ISequenceQuery<T> sequenceQuery;
